agents/llm.py
import os
from dotenv import load_dotenv
import google.generativeai as genai

# Load environment variables from .env
load_dotenv()

# Configure Gemini with API key
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

# Initialize Gemini model
model = genai.GenerativeModel("models/gemini-2.5-flash-lite")


def ask_gemini(prompt: str) -> str:
    """
    Sends a prompt to Gemini and returns the response text.
    """
    try:
        response = model.generate_content(prompt)
        return response.text
    except Exception as e:
        # return None
        return f"Error from Gemini: {str(e)}"

